Remove opencode-kimchi plugin
