+++ src/integrations/constants.ts
@@ -3,7 +3,3 @@ export const PROVIDER_NAME = "kimchi"
 export const API_KEY_ENV = "KIMCHI_API_KEY"
 export const BASE_URL = "https://llm.kimchi.dev/openai/v1"
 export const ANTHROPIC_BASE_URL = "https://llm.kimchi.dev/anthropic"
-
-export const NPM_REGISTRY_BASE_URL = "https://registry.npmjs.org"
-export const OPENCODE_PLUGIN_PACKAGE = "@kimchi-dev/opencode-kimchi"
-export const OPENCODE_PLUGIN_ARRAY_MIN_VERSION = "1.14.0"
