+++ src/integrations/openclaw.ts
@@ -5,10 +5,10 @@ import { dirname, join } from "node:path"
 import { readJson, writeFileAtomic, writeJson } from "../config/json.js"
 import type { ConfigScope } from "../config/scope.js"
 import { resolveScopePath } from "../config/scope.js"
+import { compareSemverGte } from "../utils/semver.js"
 import { API_KEY_ENV, BASE_URL, PROVIDER_NAME } from "./constants.js"
 import { findBinary } from "./detect.js"
 import { type ModelRole, resolveAllModelRoles, resolveModelRole } from "./models.js"
-import { compareSemverGte } from "./opencode.js"
 import { register } from "./registry.js"
 
 const OPENCLAW_CONFIG_PATH = "~/.openclaw/openclaw.json"
